LayoutTests/fast/mediastream/getUserMedia-deny-persistency4.html

<!DOCTYPE HTML>
<html>
    <head>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
promise_test(async (test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(false);
    await navigator.mediaDevices.getUserMedia({audio:true, video:true}).then(assert_unreached, (e) => {
        assert_equals(e.name, "NotAllowedError");
    });
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);
    const stream2 = await navigator.mediaDevices.getUserMedia({video:true});
    test.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    if (window.testRunner)
        testRunner.setUserMediaPermission(false);
    const stream3 = await navigator.mediaDevices.getUserMedia({video:true});
    test.add_cleanup(() => stream3.getTracks().forEach(track => track.stop()));

    await navigator.mediaDevices.getUserMedia({audio:true}).then(assert_unreached, (e) => {
        assert_equals(e.name, "NotAllowedError");
    });
}, "Testing same page getUserMedia deny persistency, the first denied request being audio and video, the second granted request being video");
        </script>
    </body>
</html>
